refactor: move unused enum parsers into microbenchmarks module
