+++ microbenchmarks/src/main/java/io/camunda/microbenchmarks/msgpack/enumparser/AsciiSortedStringEnumParser.java
@@ -5,8 +5,9 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.msgpack.util;
+package io.camunda.microbenchmarks.msgpack.enumparser;
 
+import io.camunda.zeebe.msgpack.util.EnumParser;
 import java.lang.reflect.Array;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
